Put setplot functions inside example scripts.
