--- examples/acoustics_2d_homogeneous/setplot.py
@@ -1,111 +1,0 @@
-
-""" 
-Set up the plot figures, axes, and items to be done for each frame.
-
-This module is imported by the plotting routines and then the
-function setplot is called to set the plot parameters.
-    
-""" 
-
-import os
-if os.path.exists('./1drad/_output'):
-    qref_dir = os.path.abspath('./1drad/_output')
-else:
-    qref_dir = None
-    print "Directory ./1drad/_output not found"
-
-
-#--------------------------
-def setplot(plotdata):
-#--------------------------
-    
-    """ 
-    Specify what is to be plotted at each frame.
-    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
-    Output: a modified version of plotdata.
-    
-    """ 
-
-
-    from clawpack.visclaw import colormaps
-
-    plotdata.clearfigures()  # clear any old figures,axes,items data
-    
-
-    # Figure for pressure
-    # -------------------
-
-    plotfigure = plotdata.new_plotfigure(name='Pressure', figno=0)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'Pressure'
-    plotaxes.scaled = True      # so aspect ratio is 1
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 0
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = True
-    plotitem.show = True       # show on plot?
-    
-
-    # Figure for scatter plot
-    # -----------------------
-    
-    plotfigure = plotdata.new_plotfigure(name='scatter', figno=1)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'Scatter plot'
-
-    # Set up for item on these axes: scatter of 2d data
-    plotitem = plotaxes.new_plotitem(plot_type='1d_from_2d_data')
-    
-    def p_vs_r(current_data):
-        # Return radius of each patch cell and p value in the cell
-        from pylab import sqrt
-        x = current_data.x
-        y = current_data.y
-        r = sqrt(x**2 + y**2)
-        q = current_data.q
-        p = q[0,:,:]
-        return r,p
-
-    plotitem.map_2d_to_1d = p_vs_r
-    plotitem.plot_var = 0
-    plotitem.plotstyle = 'ob'
-    plotitem.show = True       # show on plot?
-    
-    # Set up for item on these axes: 1d reference solution
-    #Remove 1d comparison: plotitem = emove 1d comparison:m(plot_type='1d_plot')
-    #Remove 1d comparison: plotitem.outdir = qref_dir
-    #Remove 1d comparison: plotitem.plot_var = 0
-    #Remove 1d comparison: plotitem.plotstyle = '-'
-    #Remove 1d comparison: plotitem.color = 'r'
-    #Remove 1d comparison: plotitem.kwargs = {'linewidth': 2}
-    #Remove 1d comparison: plotitem.show = True       # show on plot?
-    #Remove 1d comparison: plotaxes.afteraxes = "pylab.legend('2d data')" #Remove 1d comparison:, '1d reference solution'))"
-    
-
-    # Parameters used only when creating html and/or latex hardcopy
-    # e.g., via visclaw.frametools.printframes:
-
-    plotdata.printfigs = True                # print figures
-    plotdata.print_format = 'png'            # file format
-    plotdata.print_framenos = 'all'          # list of frames to print
-    plotdata.print_fignos = 'all'            # list of figures to print
-    plotdata.html = True                     # create html files of plots?
-    plotdata.html_homelink = '../README.html'   # pointer for top of index
-    plotdata.latex = True                    # create latex file of plots?
-    plotdata.latex_figsperline = 2           # layout of plots
-    plotdata.latex_framesperline = 1         # layout of plots
-    plotdata.latex_makepdf = False           # also run pdflatex?
-
-    return plotdata
-
-
