--- examples/acoustics_2d_variable/setplot.py
@@ -1,90 +1,0 @@
-
-""" 
-Set up the plot figures, axes, and items to be done for each frame.
-
-This module is imported by the plotting routines and then the
-function setplot is called to set the plot parameters.
-    
-""" 
-
-import os
-if os.path.exists('./1drad/_output'):
-    qref_dir = os.path.abspath('./1drad/_output')
-else:
-    qref_dir = None
-    print "Directory ./1drad/_output not found"
-
-
-#--------------------------
-def setplot(plotdata):
-#--------------------------
-    
-    """ 
-    Specify what is to be plotted at each frame.
-    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
-    Output: a modified version of plotdata.
-    
-    """ 
-
-
-    from clawpack.visclaw import colormaps
-
-    plotdata.clearfigures()  # clear any old figures,axes,items data
-    
-
-    # Figure for pressure
-    # -------------------
-
-    plotfigure = plotdata.new_plotfigure(name='Pressure', figno=0)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'Pressure'
-    plotaxes.scaled = True      # so aspect ratio is 1
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 0
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = True
-    plotitem.show = True       # show on plot?
-    plotitem.pcolor_cmin = 0.0
-    plotitem.pcolor_cmax=1.0
-    
-
-    # Figure for x-velocity plot
-    # -----------------------
-    
-    plotfigure = plotdata.new_plotfigure(name='x-Velocity', figno=1)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'u'
-
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 1
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = True
-    plotitem.show = True       # show on plot?
-    plotitem.pcolor_cmin = -0.3
-    plotitem.pcolor_cmax=   0.3
-    
-    # Parameters used only when creating html and/or latex hardcopy
-    # e.g., via visclaw.frametools.printframes:
-
-    plotdata.printfigs = True                # print figures
-    plotdata.print_format = 'png'            # file format
-    plotdata.print_framenos = 'all'          # list of frames to print
-    plotdata.print_fignos = 'all'            # list of figures to print
-    plotdata.html = True                     # create html files of plots?
-    plotdata.html_homelink = '../README.html'   # pointer for top of index
-    plotdata.latex = True                    # create latex file of plots?
-    plotdata.latex_figsperline = 2           # layout of plots
-    plotdata.latex_framesperline = 1         # layout of plots
-    plotdata.latex_makepdf = False           # also run pdflatex?
-
-    return plotdata
