--- examples/euler_2d/setplot.py
@@ -1,113 +1,0 @@
-
-""" 
-Set up the plot figures, axes, and items to be done for each frame.
-
-This module is imported by the plotting routines and then the
-function setplot is called to set the plot parameters.
-    
-""" 
-
-#--------------------------
-def setplot(plotdata):
-#--------------------------
-    
-    """ 
-    Specify what is to be plotted at each frame.
-    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
-    Output: a modified version of plotdata.
-    
-    """ 
-
-
-    from clawpack.visclaw import colormaps
-    from matplotlib import cm
-
-    plotdata.clearfigures()  # clear any old figures,axes,items data
-    
-
-    # Figure for pressure
-    # -------------------
-
-    plotfigure = plotdata.new_plotfigure(name='Density', figno=0)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'Density'
-    plotaxes.scaled = True      # so aspect ratio is 1
-    plotaxes.afteraxes = label_axes
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_schlieren')
-    #plotitem.pcolor_cmin = 0.5
-    #plotitem.pcolor_cmax=3.5
-    plotitem.plot_var = 0
-    plotitem.add_colorbar = False
-    plotitem.show = True       # show on plot?
-    
-
-    plotfigure = plotdata.new_plotfigure(name='Tracer', figno=1)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'Tracer'
-    plotaxes.scaled = True      # so aspect ratio is 1
-    plotaxes.afteraxes = label_axes
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.pcolor_cmin = 0.
-    plotitem.pcolor_cmax=1.0
-    plotitem.plot_var = 3
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = False
-    plotitem.show = True       # show on plot?
-    
-
-    plotfigure = plotdata.new_plotfigure(name='Energy', figno=2)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'Energy'
-    plotaxes.scaled = True      # so aspect ratio is 1
-    plotaxes.afteraxes = label_axes
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.pcolor_cmin = 2.
-    plotitem.pcolor_cmax=18.0
-    plotitem.plot_var = 3
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = False
-    plotitem.show = True       # show on plot?
-    
-
-
-
-    # Parameters used only when creating html and/or latex hardcopy
-    # e.g., via visclaw.frametools.printframes:
-
-    plotdata.printfigs = True                # print figures
-    plotdata.print_format = 'png'            # file format
-    plotdata.print_framenos = 'all'          # list of frames to print
-    plotdata.print_fignos = 'all'            # list of figures to print
-    plotdata.html = True                     # create html files of plots?
-    plotdata.html_homelink = '../README.html'   # pointer for top of index
-    plotdata.latex = True                    # create latex file of plots?
-    plotdata.latex_figsperline = 2           # layout of plots
-    plotdata.latex_framesperline = 1         # layout of plots
-    plotdata.latex_makepdf = False           # also run pdflatex?
-
-    return plotdata
-
-def label_axes(current_data):
-    import matplotlib.pyplot as plt
-    plt.xlabel('z')
-    plt.ylabel('r')
-    #plt.draw()
-
