+++ examples/euler_2d/shock_bubble_interaction.py
@@ -226,7 +226,88 @@ def setup(use_petsc=False,kernel_language='Fortran',solver_type='classic', outdi
     #return claw.frames[claw.num_output_times].state
     return claw
 
+#--------------------------
+def setplot(plotdata):
+#--------------------------
+    """ 
+    Specify what is to be plotted at each frame.
+    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
+    Output: a modified version of plotdata.
+    """ 
+    from clawpack.visclaw import colormaps
+    from matplotlib import cm
+
+    plotdata.clearfigures()  # clear any old figures,axes,items data
+    
+    # Figure for pressure
+    plotfigure = plotdata.new_plotfigure(name='Density', figno=0)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.xlimits = 'auto'
+    plotaxes.ylimits = 'auto'
+    plotaxes.title = 'Density'
+    plotaxes.scaled = True      # so aspect ratio is 1
+    plotaxes.afteraxes = label_axes
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='2d_schlieren')
+    #plotitem.pcolor_cmin = 0.5
+    #plotitem.pcolor_cmax=3.5
+    plotitem.plot_var = 0
+    plotitem.add_colorbar = False
+    plotitem.show = True       # show on plot?
+    
+
+    plotfigure = plotdata.new_plotfigure(name='Tracer', figno=1)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.xlimits = 'auto'
+    plotaxes.ylimits = 'auto'
+    plotaxes.title = 'Tracer'
+    plotaxes.scaled = True      # so aspect ratio is 1
+    plotaxes.afteraxes = label_axes
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
+    plotitem.pcolor_cmin = 0.
+    plotitem.pcolor_cmax=1.0
+    plotitem.plot_var = 3
+    plotitem.pcolor_cmap = colormaps.yellow_red_blue
+    plotitem.add_colorbar = False
+    plotitem.show = True       # show on plot?
+    
+
+    plotfigure = plotdata.new_plotfigure(name='Energy', figno=2)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.xlimits = 'auto'
+    plotaxes.ylimits = 'auto'
+    plotaxes.title = 'Energy'
+    plotaxes.scaled = True      # so aspect ratio is 1
+    plotaxes.afteraxes = label_axes
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
+    plotitem.pcolor_cmin = 2.
+    plotitem.pcolor_cmax=18.0
+    plotitem.plot_var = 3
+    plotitem.pcolor_cmap = colormaps.yellow_red_blue
+    plotitem.add_colorbar = False
+    plotitem.show = True       # show on plot?
+    
+    return plotdata
+
+def label_axes(current_data):
+    import matplotlib.pyplot as plt
+    plt.xlabel('z')
+    plt.ylabel('r')
+    #plt.draw()
+    
+
 if __name__=="__main__":
     from clawpack.pyclaw.util import run_app_from_main
-    output = run_app_from_main(setup)
+    output = run_app_from_main(setup,setplot)
