+++ examples/euler_2d/shockbubble_scipy.py
@@ -223,4 +223,5 @@ def shockbubble(use_petsc=False,outdir='./_output',solver_type='classic'):
 
 if __name__=="__main__":
     from clawpack.pyclaw.util import run_app_from_main
-    output = run_app_from_main(shockbubble)
+    from clawpack.pyclaw.examples import *
+    output = run_app_from_main(shockbubble,shock_bubble_interaction.setplot)
