--- examples/shallow_2d/setplot.py
@@ -1,122 +1,0 @@
-
-""" 
-Set up the plot figures, axes, and items to be done for each frame.
-
-This module is imported by the plotting routines and then the
-function setplot is called to set the plot parameters.
-    
-""" 
-
-#--------------------------
-def setplot(plotdata):
-#--------------------------
-    
-    """ 
-    Specify what is to be plotted at each frame.
-    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
-    Output: a modified version of plotdata.
-    
-    """ 
-
-
-    from clawpack.visclaw import colormaps
-
-    plotdata.clearfigures()  # clear any old figures,axes,items data
-    
-
-    # Figure for q[0]
-    plotfigure = plotdata.new_plotfigure(name='Water height', figno=0)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = [-2.5, 2.5]
-    plotaxes.ylimits = [-2.5, 2.5]
-    plotaxes.title = 'Water height'
-    plotaxes.scaled = True
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 0
-    plotitem.pcolor_cmap = colormaps.red_yellow_blue
-    plotitem.pcolor_cmin = 0.5
-    plotitem.pcolor_cmax = 1.5
-    plotitem.add_colorbar = True
-    plotitem.show = True       # show on plot?
-    
-
-    # Scatter plot of q[0]
-    plotfigure = plotdata.new_plotfigure(name='Scatter plot of h', figno=1)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = [0., 2.5]
-    plotaxes.ylimits = [0., 2.1]
-    plotaxes.title = 'Scatter plot of h'
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='1d_from_2d_data')
-    plotitem.plot_var = 0
-    def q_vs_radius(current_data):
-        from numpy import sqrt
-        x = current_data.x
-        y = current_data.y
-        r = sqrt(x**2 + y**2)
-        q = current_data.q[0,:,:]
-        return r,q
-    plotitem.map_2d_to_1d = q_vs_radius
-    plotitem.plotstyle = 'o'
-
-
-    # Figure for q[1]
-    plotfigure = plotdata.new_plotfigure(name='Momentum in x direction', figno=2)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = [-2.5, 2.5]
-    plotaxes.ylimits = [-2.5, 2.5]
-    plotaxes.title = 'Momentum in x direction'
-    plotaxes.scaled = True
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 1
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = True
-    plotitem.show = False       # show on plot?
-    
-
-    # Figure for q[2]
-    plotfigure = plotdata.new_plotfigure(name='Momentum in y direction', figno=3)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = [-2.5, 2.5]
-    plotaxes.ylimits = [-2.5, 2.5]
-    plotaxes.title = 'Momentum in y direction'
-    plotaxes.scaled = True
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 2
-    plotitem.pcolor_cmap = colormaps.yellow_red_blue
-    plotitem.add_colorbar = True
-    plotitem.show = False       # show on plot?
-    
-
-    # Parameters used only when creating html and/or latex hardcopy
-    # e.g., via visclaw.frametools.printframes:
-
-    plotdata.printfigs = True                # print figures
-    plotdata.print_format = 'png'            # file format
-    plotdata.print_framenos = 'all'          # list of frames to print
-    plotdata.print_fignos = 'all'            # list of figures to print
-    plotdata.html = True                     # create html files of plots?
-    plotdata.html_homelink = '../README.html'   # pointer for top of index
-    plotdata.latex = True                    # create latex file of plots?
-    plotdata.latex_figsperline = 2           # layout of plots
-    plotdata.latex_framesperline = 1         # layout of plots
-    plotdata.latex_makepdf = False           # also run pdflatex?
-
-    return plotdata
-
-
