--- examples/stegoton_1d/setplot.py
@@ -1,92 +1,0 @@
-
-""" 
-Set up the plot figures, axes, and items to be done for each frame.
-
-This module is imported by the plotting routines and then the
-function setplot is called to set the plot parameters.
-    
-""" 
-
-#--------------------------
-def setplot(plotdata):
-#--------------------------
-    
-    """ 
-    Specify what is to be plotted at each frame.
-    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
-    Output: a modified version of plotdata.
-    
-    """ 
-
-
-    plotdata.clearfigures()  # clear any old figures,axes,items data
-
-    # Figure for q[0]
-    plotfigure = plotdata.new_plotfigure(name='Stress', figno=1)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    #plotaxes.xlimits = [0.,150.]
-    #plotaxes.ylimits = [-.2,1.0]
-    plotaxes.title = 'Stress'
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='1d_plot')
-    plotitem.plot_var = stress
-    plotitem.plotstyle = '-'
-    plotitem.color = 'b'
-    plotitem.show = True       # show on plot?
-    plotitem.kwargs = {'linewidth':2,'markersize':5}
-    
-
-
-    # Figure for q[1]
-    plotfigure = plotdata.new_plotfigure(name='Velocity', figno=2)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = [-.5,1.1]
-    plotaxes.title = 'Velocity'
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='1d_plot')
-    plotitem.plot_var = velocity
-    plotitem.plotstyle = '-'
-    plotitem.color = 'b'
-    plotitem.show = True       # show on plot?
-    plotitem.kwargs = {'linewidth':3,'markersize':5}
-    
-
-    # Parameters used only when creating html and/or latex hardcopy
-    # e.g., via visclaw.frametools.printframes:
-
-    plotdata.printfigs = True                # print figures
-    plotdata.print_format = 'png'            # file format
-    plotdata.print_framenos = 'all'          # list of frames to print
-    plotdata.print_fignos = 'all'            # list of figures to print
-    plotdata.html = True                     # create html files of plots?
-    plotdata.html_homelink = '../README.html'
-    plotdata.latex = True                    # create latex file of plots?
-    plotdata.latex_figsperline = 2           # layout of plots
-    plotdata.latex_framesperline = 1         # layout of plots
-    plotdata.latex_makepdf = False           # also run pdflatex?
-
-    return plotdata
-
- 
-def velocity(current_data):
-    """Compute velocity from strain and momentum"""
-    from stegoton import setaux
-    aux=setaux(current_data.x,rhoB=4,KB=4)
-    velocity = current_data.q[1,:]/aux[0,:]
-    return velocity
-
-def stress(current_data):
-    """Compute stress from strain and momentum"""
-    from stegoton import setaux
-    from clawpack.riemann.nonlinear_elasticity_1D_py import sigma 
-    aux=setaux(current_data.x)
-    epsilon = current_data.q[0,:]
-    stress = sigma(epsilon,aux[1,:])
-    return stress
