+++ examples/stegoton_1d/stegoton.py
@@ -166,6 +166,67 @@ def setup(use_petsc=0,kernel_language='Fortran',solver_type='classic',outdir='./
     return claw
 
 
+#--------------------------
+def setplot(plotdata):
+#--------------------------
+    """ 
+    Specify what is to be plotted at each frame.
+    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
+    Output: a modified version of plotdata.
+    """ 
+    plotdata.clearfigures()  # clear any old figures,axes,items data
+
+    # Figure for q[0]
+    plotfigure = plotdata.new_plotfigure(name='Stress', figno=1)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.title = 'Stress'
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='1d_plot')
+    plotitem.plot_var = stress
+    plotitem.plotstyle = '-'
+    plotitem.color = 'b'
+    plotitem.show = True       # show on plot?
+    plotitem.kwargs = {'linewidth':2,'markersize':5}
+    
+    # Figure for q[1]
+    plotfigure = plotdata.new_plotfigure(name='Velocity', figno=2)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.xlimits = 'auto'
+    plotaxes.ylimits = [-.5,1.1]
+    plotaxes.title = 'Velocity'
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='1d_plot')
+    plotitem.plot_var = velocity
+    plotitem.plotstyle = '-'
+    plotitem.color = 'b'
+    plotitem.show = True       # show on plot?
+    plotitem.kwargs = {'linewidth':3,'markersize':5}
+    
+    return plotdata
+
+ 
+def velocity(current_data):
+    """Compute velocity from strain and momentum"""
+    from stegoton import setaux
+    aux=setaux(current_data.x,rhoB=4,KB=4)
+    velocity = current_data.q[1,:]/aux[0,:]
+    return velocity
+
+def stress(current_data):
+    """Compute stress from strain and momentum"""
+    from stegoton import setaux
+    from clawpack.riemann.nonlinear_elasticity_1D_py import sigma 
+    aux=setaux(current_data.x)
+    epsilon = current_data.q[0,:]
+    stress = sigma(epsilon,aux[1,:])
+    return stress
+
 if __name__=="__main__":
     from clawpack.pyclaw.util import run_app_from_main
-    output = run_app_from_main(setup)
+    output = run_app_from_main(setup,setplot)
