+++ src/pyclaw/plot.py
@@ -4,9 +4,11 @@ import os
 import sys
 import types
 
-def plot(setplot_path=None,outdir="./_output",plotdir=None,htmlplot=False,iplot=True,
+def plot(setplot=None,outdir="./_output",plotdir=None,htmlplot=False,iplot=True,
          file_format='ascii',**plot_kargs):
-    r""""""
+    r"""
+    setplot can be a function or a path to a file.
+    """
     
     # Construct a plot directory if not provided
     if plotdir is None:
@@ -17,23 +19,24 @@ def plot(setplot_path=None,outdir="./_output",plotdir=None,htmlplot=False,iplot=
     
     if htmlplot or iplot:
         # Grab and import the setplot function
-        # Use local setplot if available
-        if setplot_path is None:
-            local_setplot_path = os.path.join(os.getcwd(),'setplot.py')
-            if os.path.exists(local_setplot_path):
-                setplot_path = local_setplot_path
+        if not isinstance(setplot,types.FunctionType):
+            # Use local setplot if available
+            if setplot is None:
+                local_setplot_path = os.path.join(os.getcwd(),'setplot.py')
+                if os.path.exists(local_setplot_path):
+                    setplot = local_setplot_path
 
-        # If setplot_path is still None then we import the default setplot
-        if setplot_path is None:
-            import clawpack.visclaw.setplot_default as setplot_module
-        else:
-            path = os.path.abspath(os.path.expandvars(os.path.expanduser(setplot_path)))
-            setplot_module_dir = os.path.dirname(path)
-            setplot_module_name = os.path.splitext(os.path.basename(setplot_path))[0]
-            sys.path.insert(0,setplot_module_dir)
-            setplot_module = __import__(setplot_module_name)
-        reload(setplot_module)
-        setplot = lambda plotdata:setplot_module.setplot(plotdata,**plot_kargs)
+            # If setplot is still None then we import the default setplot
+            if setplot is None:
+                import clawpack.visclaw.setplot_default as setplot_module
+            else:
+                path = os.path.abspath(os.path.expandvars(os.path.expanduser(setplot)))
+                setplot_module_dir = os.path.dirname(path)
+                setplot_module_name = os.path.splitext(os.path.basename(setplot))[0]
+                sys.path.insert(0,setplot_module_dir)
+                setplot_module = __import__(setplot_module_name)
+            reload(setplot_module)
+            setplot = lambda plotdata:setplot_module.setplot(plotdata,**plot_kargs)
         
         if not isinstance(setplot,types.FunctionType):
             raise ImportError("Failed to import %s.setplot" % setplot_module_name)
@@ -52,14 +55,14 @@ def plot(setplot_path=None,outdir="./_output",plotdir=None,htmlplot=False,iplot=
         
 
 # These now just point to the above more generic function
-def interactive_plot(outdir='./_output',file_format='ascii'):
+def interactive_plot(outdir='./_output',file_format='ascii',setplot=None):
     """
     Convenience function for launching an interactive plotting session.
     """
-    plot(outdir=outdir,file_format=file_format,iplot=True,htmlplot=False)
+    plot(setplot,outdir=outdir,file_format=file_format,iplot=True,htmlplot=False)
 
-def html_plot(outdir='./_output',file_format='ascii'):
+def html_plot(outdir='./_output',file_format='ascii',setplot=None):
     """
     Convenience function for creating html page with plots.
     """
-    plot(outdir=outdir,file_format=file_format,htmlplot=True,iplot=False)
+    plot(setplot,outdir=outdir,file_format=file_format,htmlplot=True,iplot=False)
